add booking and fix tour

+++ BE/src/main/java/com/devteria/identityservice/dto/request/TourRequest.java
@@ -11,7 +11,7 @@ public class TourRequest {
     private String tourBanner;
     private String description;
     private Double price;
-    private Integer managerId; // managerId để ánh xạ với User entity
+    private Long managerId; // managerId để ánh xạ với User entity
     private String companyName;
     private String location;
     private Integer totalTicket;
